+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # CHANGELOG
 
-## 1.8.3 (01.02.2021)
+## 1.8.4 (22.02.2021)
 - improvment
     - cli test covered
+
+## 1.8.3 (22.02.2021)
+- improvement
+    - add default docker user definition
+    - add unittests for general command lib
 
 ## 1.8.2 (31.01.2021)
 - improvment